refactored Controller into GUIController, pt1, AB#1400
