+++ src/main/java/com/lastcruise/Main.java
@@ -1,22 +1,24 @@
 package com.lastcruise;
 
-import com.lastcruise.controller.Controller;
 import com.lastcruise.controller.GUIController;
+import com.lastcruise.model.AllSounds;
+import com.lastcruise.model.Music;
 import java.io.IOException;
+import java.net.URL;
 
 public class Main {
 
   public static void main(String[] args) throws IOException, InterruptedException {
 
-      GUIController gui = new GUIController();
-      // creates guiController instance in Controller, Prompts the user to start game or quit
-      gui.gameSetUp();
+    GUIController gui = new GUIController();
+    // creates guiController instance in Controller, Prompts the user to start game or quit
+    gui.gameSetUp();
 
-      boolean runGame = true;
-      // TODO: Goal to enter while loop and not change too much logic here
-//      URL backgroundMusic = Main.class.getResource(AllSounds.ALL_SOUNDS.get("main"));
-//      Music.runAudio(backgroundMusic);
-//
+//    boolean runGame = gui.getKeepPlaying();
+    // TODO: Goal to enter while loop and not change too much logic here
+//    URL backgroundMusic = Main.class.getResource(AllSounds.ALL_SOUNDS.get("main"));
+//    Music.runAudio(backgroundMusic);
+
 //    while (runGame) {
 //      runGame = controller.getCommand();
 //      controller.updateView();
