+++ src/main/java/com/lastcruise/controller/GUIController.java
@@ -1,7 +1,17 @@
 package com.lastcruise.controller;
 
+import com.lastcruise.model.AllSounds;
+import com.lastcruise.model.Commands;
+import com.lastcruise.model.CraftingLocation;
 import com.lastcruise.model.Game;
+import com.lastcruise.model.GameMap.InvalidLocationException;
+import com.lastcruise.model.Inventory.InventoryEmptyException;
+import com.lastcruise.model.Music;
 import com.lastcruise.model.Player;
+import com.lastcruise.model.Player.ItemNotEdibleException;
+import com.lastcruise.model.Player.NoEnoughStaminaException;
+import com.lastcruise.model.PuzzleClient;
+import com.lastcruise.model.SoundEffect;
 import com.lastcruise.view.GameScreen;
 import com.lastcruise.view.PreludeScreen;
 import com.lastcruise.view.TitleScreen;
@@ -10,6 +20,8 @@ import java.awt.Dimension;
 import java.awt.Image;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.io.IOException;
+import java.net.URL;
 import javax.swing.ImageIcon;
 import javax.swing.JFrame;
 import javax.swing.JLayeredPane;
@@ -18,27 +30,41 @@ import javax.swing.Timer;
 
 public class GUIController {
 
+  // Primary Window Frame
   private JFrame mainFrame;
+
+  // Changing Panels
   private final TitleScreen titleScreen = new TitleScreen();
   private PreludeScreen intermission = new PreludeScreen();
   // Use callback instead of static methods
   private GameScreen mainGameScreen = new GameScreen();
-  static View view = new View();
-  private Controller controller = new Controller();
-  private Game game;
 
+  // Controller Related Fields
+  private final View view = new View();
+  PuzzleClient puzzleClient = new PuzzleClient();
+  private String name;
+  private String message;
+  private final GameLoader gameLoader = new GameLoader();
+  private Game game;
+  private Controller controller = new Controller();
+  private Boolean keepPlaying = true;
+
+  // CONSTRUCTOR
   public GUIController() {
+    // CREATE GAME WINDOW/FRAME
     mainFrame = new JFrame("The Last Cruise");
     mainFrame.setPreferredSize(new Dimension(1500, 800));
     mainFrame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
     mainFrame.setResizable(false);
     ImageIcon image = new ImageIcon(getClass().getClassLoader().getResource("images/logo.jpg"));
     mainFrame.setIconImage(image.getImage());
+    // LOAD TITLE SCREEN
     loadTitleScreen();
+    // ACTION CALLBACK TO PROCESS MAIN GAME COMMANDS
     mainGameScreen.setActionCallback((commands) -> {
       try {
-        controller.processCommand(commands, game);
-        // make ure to attach with each event listener to update view
+        processCommand(commands);
+        // make sure to attach with each event listener to update view
         updateView();
       } catch (InterruptedException e) {
         throw new RuntimeException(e);
@@ -96,8 +122,8 @@ public class GUIController {
     updateView();
   }
 
-  private void setGameText(String message) {
-    mainGameScreen.getDialogueTextArea().setText(message);
+  private void addGameText(String message) {
+    mainGameScreen.getDialogueTextArea().append(message);
   }
 
   public void updateView() {
@@ -107,12 +133,18 @@ public class GUIController {
     String stamina = game.getPlayerStamina();
     String locationDesc = game.getCurrentLocationDesc();
     String locationItems = game.getCurrentLocationItems().keySet().toString();
+
     // update GUI View with current location and Stamina
     mainGameScreen.getLocationLabel().setText(location);
     mainGameScreen.getStaminaLabel().setText(stamina);
-    updateLocationImg(location);
-  }
+    mainGameScreen.getDialogueTextArea().setText(locationDesc);
 
+    // Updates GameScreen Location image
+    updateLocationImg(location);
+
+    // append additional game messages to the scroll screen
+    addGameText(message);
+  }
 
   public void updateLocationImg(String location) {
     ImageIcon icon = new ImageIcon(getClass().getClassLoader().getResource("images/" + location +".jpg"));
@@ -122,6 +154,216 @@ public class GUIController {
     mainGameScreen.getBgImgLabel().setIcon(icon);
   }
 
+  public void processCommand(String[] command) throws InterruptedException {
+    Commands c = Commands.valueOf(command[0].toUpperCase());
+
+    switch (c) {
+      //---- GO -------//
+      case GO: {
+        try {
+          game.moveLocation(command);
+          // TODO: Adjust PIT to print message elsewhere once panel is decided
+          if (game.getCurrentLocationName().equals("PIT")) {
+
+            URL fallSoundUrl = getClass().getResource(
+                AllSounds.ALL_SOUNDS.get("pitFall"));
+            SoundEffect.runAudio(fallSoundUrl);
+
+            message = view.pitFallPrompt();
+            updateView();
+//            updateLocationTimer();
+            message = view.puzzleMessagePrompt();
+            updateView();
+
+            if (puzzleClient.puzzleGenerator()) {
+              message = view.solvedPuzzleMessage();
+            } else {
+              message = view.unSolvedPuzzleMessage();
+              updateView();
+              puzzleClient.puzzlePunishment();
+              message = view.pitFallEscapePrompt();
+            }
+          } else {
+            URL runSoundUrl = getClass().getResource(
+                AllSounds.ALL_SOUNDS.get("run"));
+            SoundEffect.runAudio(runSoundUrl);
+          }
+
+        } catch (InvalidLocationException e) {
+          message = view.getInvalidLocationMessage();
+
+        } catch (InterruptedException e) {
+          throw new RuntimeException(e);
+
+        } catch (NoEnoughStaminaException e) {
+          message = view.getNoStaminaToMove();
+
+        }
+        break;
+      }
+      //---- HELP -------//
+      case HELP: {
+        message = view.getHelpCommands();
+        break;
+      }
+      //---- INSPECT -------//
+      case INSPECT: {
+        if (game.inspectItem(command) != null) {
+          message = view.getItemDescription(game.inspectItem(command));
+        }
+        break;
+      }
+      //---- GRAB -------//
+      case PICKUP:
+      case TAKE:
+      case GRAB: {
+        var currentLocationInventory = game.getCurrentLocationInventory();
+        var playerInventory = game.getPlayerInventory();
+        // GRABBING LOG
+        if (command[1].equals("log") && !playerInventory.getInventory()
+            .containsKey("machete")) {
+          message = view.cantGrabItem();
+        } else {
+          try {
+            game.transferItemFromTo(currentLocationInventory, playerInventory,
+                command[1]);
+            URL grabSoundUrl = getClass().getResource(
+                AllSounds.ALL_SOUNDS.get("pickup"));
+            SoundEffect.runAudio(grabSoundUrl);
+          } catch (InventoryEmptyException e) {
+            message = view.getInvalidItemMessage();
+          } catch (NoEnoughStaminaException e) {
+            message = view.getNoPickUpStamina();
+          }
+
+        }
+        break;
+      }
+      //--- DROP ------//
+      case DROP: {
+        var currentLocationInventory = game.getCurrentLocationInventory();
+        var playerInventory = game.getPlayerInventory();
+        try {
+          game.transferItemFromTo(playerInventory,
+              currentLocationInventory,
+              command[1]);
+          URL dropSoundUrl = getClass().getResource(AllSounds.ALL_SOUNDS.get("drop"));
+          SoundEffect.runAudio(dropSoundUrl);
+        } catch (InventoryEmptyException e) {
+          message = view.getInvalidItemMessage();
+        } catch (NoEnoughStaminaException e) {
+          message = view.getNoDropStamina();
+        }
+        break;
+      }
+      //--- CRAFT ---//
+      case BUILD:
+      case CRAFT: {
+        if (command[1].equals("raft")) {
+          if (game.getCurrentLocation() instanceof CraftingLocation) {
+            //Craft raft
+            if (game.craftRaft()) {
+              message = view.getSuccessfulRaftBuildMessage();
+
+            } else {
+              message = view.getNotSuccessfulRaftBuildMessage();
+            }
+
+          } else {
+            message = view.getNotInRaftLocationBuildMessage();
+          }
+
+        } else {
+          message = view.getItemNotCraftable();
+        }
+        break;
+      }
+      case EAT: {
+        try {
+          game.eatItem(command[1]);
+          message = view.getEating();
+          URL eatSoundUrl = getClass().getResource(AllSounds.ALL_SOUNDS.get("eat"));
+          SoundEffect.runAudio(eatSoundUrl);
+        } catch (InventoryEmptyException e) {
+          message = view.getInvalidItemMessage();
+        } catch (ItemNotEdibleException e) {
+          message = view.getCantEatThat();
+        }
+        break;
+      }
+      case SLEEP: {
+        game.playerSleep();
+        message = view.getSleeping();
+        Thread.sleep(4000);
+        break;
+      }
+      case ESCAPE: {
+        if (game.getCurrentLocationItems().containsKey("raft")) {
+          message = view.getYouWonMessage();
+          keepPlaying = false;
+        } else {
+          message = view.getCantEscape();
+        }
+        break;
+      }
+      case QUIT: {
+        keepPlaying = false;
+        break;
+      }
+      // MUSIC and SOUND EFFECTS CONTROLS
+      case VOLUME:
+      case MUSIC: {
+        if (command[1].equals("up")) {
+          Music.increaseMusic();
+          break;
+        } else if (command[1].equals("down")) {
+          Music.decreaseMusic();
+          break;
+        } else if (command[1].equals("off") || command[1].equals("mute") || command[1].equals("stop")) {
+          Music.muteMusic();
+          break;
+        } else if (command[1].equals("on") || command[1].equals("unmute")) {
+          Music.unMuteMusic();
+          break;
+        }
+      }
+      case SOUND: {
+        if (command[1].equals("up")) {
+          SoundEffect.increaseFxVolume();
+          break;
+        } else if (command[1].equals("down")) {
+          SoundEffect.decreaseFxVolume();
+          break;
+        } else if (command[1].equals("off") || command[1].equals("mute")) {
+          SoundEffect.muteSoundFx();
+          break;
+        } else if (command[1].equals("on") || command[1].equals("unmute")) {
+          SoundEffect.unMuteSoundFx();
+          break;
+        }
+      }
+
+      // SAVING THE GAME
+      case SAVE: {
+        if (command[1].equals("game")) {
+          try {
+            gameLoader.saveGame(game);
+            message = view.getGameSaved();
+          } catch (IOException e) {
+            message = view.getGameSaveFailed();
+          }
+          break;
+        }
+      }
+      default:
+        message = view.getInvalidCommandMessage();
+        break;
+    }
+  }
+
+  public Boolean getKeepPlaying() {
+    return keepPlaying;
+  }
 
   public JFrame getMainFrame() {
     return mainFrame;
